put the list of parameters in separate file
